+++ widgets/EntryParticipants.php
@@ -63,4 +63,20 @@ class EntryParticipants extends Widget
                      ->icon($participantSate === $state ? 'fa-check' : null)
                      ->action('calendar.respond', Url::toEntryRespond($calendarEntry, $state));
     }
+
+    public static function inviteButton(CalendarEntry $calendarEntry): string
+    {
+        if ($calendarEntry->participation_mode !== CalendarEntry::PARTICIPATION_MODE_INVITE) {
+            return '';
+        }
+
+        if (!$calendarEntry->isOwner()) {
+            return '';
+        }
+
+        return Button::asLink(Yii::t('CalendarModule.views_entry_view', 'Invite participants'), Url::toInviteParticipants($calendarEntry))
+            ->cssClass('btn btn-primary')->sm()
+            ->icon('fa-users')
+            ->options(['data-target' => '#globalModal']);
+    }
 }